Enable capability to consume ROSA masterbranch (#108)
Right now the cli option in ROSA to deploy OVN clusters is not
in a tagged build. In the future, this might be the case for other
features. It would be nice to have the ability to consume master branch
ROSA. This commit involves changes in the Dockerfile for the
airflow-managed-services image because we would need go in the image and
right now with multi-stage builds we are only using the golang image for
building osde2e binary and then copying  it off to the airflow base
image.

This commit also removes all references to osde2d since its no longer
required after ROSA supports OVNKubernetes installs.

Signed-off-by: Sai Sindhur Malleni <smalleni@redhat.com>