Run non OCP workloads from airflow
Some workloads doesn't need OCP environment. For example, ocm-api-load
which tests OCM server can be run as a standalone application on a
jump host. Any tool which can be run on a jump host, however to be
triggered using airflow can use this framework. With this we can
avoid multiple point of failures like OCP installation,
benchmark-operator, e2e if the tool is not dependent on them.

nocp_benchmarks.md doc added explaining
1) Assumptions in this approach
2) How to add new workload

ocm benchmark script
1) creates new AWS key for OsdCcsAdmin user which is required for
   OCM testing
2) runs ocm-api-load test for all endpoints
3) pushes results to snappy server
4) scrap metrics from AppSRE servers and push to our ES