Dropping SDN, adding 4.13
Trying to cleanup some of the dags, since the defaults SDN is now OVNK,
no need to carry SDN going forward. If we need to spot check, someone
can create a branch and run.

Need to confirm with the Managed Services team this is cool to do for
ROSA and others.

Signed-off-by: Joe Talerico <jtaleric@redhat.com>